requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
aiosqlite==0.20.0
Pillow==10.3.0
pytesseract==0.3.10
slowapi==0.1.9
loguru==0.7.2
httpx==0.27.0
